pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mobileloadx"
version = "1.0.0"
description = "Framework de Performance Testing para Apps Mobile - Simule múltiplos usuários simultâneos e colete métricas detalhadas"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name = "MobileLoadX Team", email = "team@mobileloadx.com" }
]
keywords = [
    "mobile",
    "testing",
    "performance",
    "load-testing",
    "appium",
    "android",
    "ios"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: Portuguese (Brazilian)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    "Appium-Python-Client>=2.11.0",
    "selenium>=4.15.0",
    "PyYAML>=6.0",
    "click>=8.1.0",
    "colorama>=0.4.6",
    "psutil>=5.9.0",
    "requests>=2.31.0",
    "Jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-timeout>=2.2.0",
    "black>=23.0.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "isort>=5.13.0",
    "bandit>=1.7.5",
]

[project.urls]
Homepage = "https://github.com/mobileloadx/mobileloadx"
Documentation = "https://github.com/mobileloadx/mobileloadx/wiki"
Repository = "https://github.com/mobileloadx/mobileloadx.git"
Issues = "https://github.com/mobileloadx/mobileloadx/issues"

[project.scripts]
mobileloadx = "mobileloadx.cli:main"

[tool.setuptools]
packages = ["mobileloadx"]

[tool.black]
line-length = 120
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 120
py_version = 39
multi_line_mode = 3
include_trailing_comma = true

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", "build", "dist", ".venv"]

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_return_any = true

[tool.coverage.run]
source = ["mobileloadx"]
omit = [
    "*/tests/*",
    "*/setup.py",
    "*/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --strict-markers --tb=short"
markers = [
    "unit: Unit test",
    "integration: Integration test",
    "slow: Slow test",
    "cli: CLI test",
    "metrics: Metrics test",
    "config: Configuration test",
]

[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.1.0",
]
